convert class component into functional component and created separate files
